fix: move clock functionality to own file
