+++ packages/apps/demo/src/components/Clock.js
@@ -1,0 +1,31 @@
+import { TextBox } from "@lightningjs/ui-components"
+
+export default class Clock extends TextBox {
+  _construct() {
+    super._construct();
+    this._startTime();
+  }
+
+  _update() {
+    this.style.textStyle = this.theme.typography.headline1;
+    super._update();
+  }
+
+  _startTime() {
+    const today = new Date();
+    let h = today.getHours();
+    let m = today.getMinutes();
+    let s = today.getSeconds();
+    m = this._checkTime(m);
+    s = this._checkTime(s);
+    this.content = h + ':' + m;
+    setTimeout(this._startTime.bind(this), 1000);
+  }
+
+  _checkTime(i) {
+    if (i < 10) {
+      i = '0' + i;
+    } // add zero in front of numbers < 10
+    return i;
+  }
+}
