+++ packages/apps/demo/src/components/index.js
@@ -18,4 +18,5 @@
 
 export { default as ThemeSwitcher } from './ThemeSwitcher';
 export { default as BackgroundSwitcher } from './BackgroundSwitcher';
-export { default as TopNav } from './TopNav'; 
+export { default as TopNav } from './TopNav';
+export { default as Clock } from './Clock';
